requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
opencv-python-headless==4.9.0.80
pillow==10.3.0
numpy==1.26.4
tensorflow==2.16.1
pydantic==2.7.4
jinja2==3.1.4
aiofiles==23.2.1
